taxi_bigdata_project/Apache_spark_folder/requirements.txt
apache-airflow-providers-snowflake
snowflake-connector-python
snowflake-sqlalchemy
pandas
sqlalchemy
pyspark
python-dotenv
pyarrow